refactor: Move CodeUsageNode class into its own file
